refactor: streamline JSON Schema normalization and parser integration
