+++ apps/craft/tests/integrations/fixtures/raw-boolean-schema.json
